+++ source/BrutePartitioning.cpp
@@ -5,11 +5,11 @@ void BrutePartitioning::Build(std::vector<const Primitive*> primitives) {
     this->primitives = primitives;
 }
 
-float BrutePartitioning::Intersect(const Ray& ray, const Primitive* &primitive, float tmin, float tmax, bool returnPrimitive=true) const {
+double BrutePartitioning::Intersect(const Ray& ray, const Primitive* &primitive, double tmin, double tmax, bool returnPrimitive=true) const {
     bool found = false;
-    float mint = inf;
+    double mint = inf;
     for(auto p : primitives) {
-        float t = p->Intersect(ray);
+        double t = p->Intersect(ray);
         if(t > -inf && t < mint && t >= tmin && t <= tmax) {
             if(!returnPrimitive)
                 return true;
