+++ source/Color.h
@@ -10,7 +10,7 @@
 class Color
 {
 public:
-    Color(float, float, float);
+    Color(double, double, double);
     //Color(const Vector3d&);
     Color(int);
     Color();
@@ -18,41 +18,41 @@ public:
     ~Color();
     Color operator+(const Color& v) const;
     Color operator-(const Color& v) const;
-    Color operator*(float) const;
-    Color operator/(float) const;
+    Color operator*(double) const;
+    Color operator/(double) const;
     Color operator*(int) const;
     Color operator/(int) const;
     Color operator*(const Color& v) const;
     Color operator+=(const Color& c);
     Color operator*=(const Color& c);
-    Color operator/=(float t);
-    Color operator*=(float t);
-    float& operator[](int);
+    Color operator/=(double t);
+    Color operator*=(double t);
+    double& operator[](int);
     bool operator!() const;
 
     bool IsValid() const;
 
-    void SetLuminance(float L);
-    float GetLuminance() const;
-    float GetIntensity() const;
-    float GetAverage() const;
-    float GetMax() const;
-    float GetSum() const;
+    void SetLuminance(double L);
+    double GetLuminance() const;
+    double GetIntensity() const;
+    double GetAverage() const;
+    double GetMax() const;
+    double GetSum() const;
 
-    float r;
-    float g;
-    float b;
+    double r;
+    double g;
+    double b;
 
     const static Color Identity;
     const static Color Black;
 
 private:
-    Color(float); // "Color(float) = delete" in public if using more C++11
+    Color(double); // "Color(double) = delete" in public if using more C++11
 };                // compliant compiler. This is to avoid accidentally 
-                  // initializing a color with a float, which never makes
+                  // initializing a color with a double, which never makes
                   // sense. Maybe I should remove the int constructor too
 
-Color operator*(float, const Color&);
+Color operator*(double, const Color&);
 Color operator*(int, const Color&);
 std::ostream& operator << (std::ostream& s, const Color& v);
