+++ source/CsgCuboid.h
@@ -9,21 +9,21 @@ class CsgCuboid : public CsgObject
 {
 public:
     CsgCuboid(const Vector3d& position, const Vector3d& x, const Vector3d& y, 
-              float a, float b, float c);
+              double a, double b, double c);
 
     bool Intersect(const Ray& ray, std::vector<CsgHit>& intersects) const;
 
     virtual BoundingBox GetBoundingBox() const;
     virtual bool GetClippedBoundingBox(const BoundingBox& clipbox, BoundingBox& resultbox) const;
 
-    virtual float Intersect(const Ray& ray) const;
+    virtual double Intersect(const Ray& ray) const;
     virtual bool GenerateIntersectionInfo(const Ray& ray, IntersectionInfo& info) const;
 
     void SetMaterial(Material* material);
     Material* GetMaterial() const;
 
     void Translate(const Vector3d& direction);
-    void Rotate(const Vector3d& axis, float angle);
+    void Rotate(const Vector3d& axis, double angle);
 
     void AddToScene(Scene& scene);
     void Save(Bytestream& stream) const;
@@ -32,10 +32,10 @@ public:
     virtual std::unique_ptr<CsgObject> Clone();
 private:
     void Precalculate();
-    bool SlabsTest(const Ray& inRay, float& tNear, float& tFar, int& axisNear, int& axisFar, int& sideNear, int& sideFar) const;
+    bool SlabsTest(const Ray& inRay, double& tNear, double& tFar, int& axisNear, int& axisFar, int& sideNear, int& sideFar) const;
 
     Vector3d pos_, x_, y_, z_;
-    float a_, b_, c_;
+    double a_, b_, c_;
 
     Vector3d invMatU_, invMatV_, invMatW_;
 };
