+++ source/CsgCylinder.h
@@ -9,21 +9,21 @@ class CsgCylinder : public CsgObject
 {
 public:
     CsgCylinder(Vector3d& position, Vector3d& dir, 
-                float length, float radius);
+                double length, double radius);
 
     bool Intersect(const Ray& ray, hits& intersects) const;
 
     virtual BoundingBox GetBoundingBox() const;
     virtual bool GetClippedBoundingBox(const BoundingBox& clipbox, BoundingBox& resultbox) const;
 
-    virtual float Intersect(const Ray& ray) const;
+    virtual double Intersect(const Ray& ray) const;
     virtual bool GenerateIntersectionInfo(const Ray& ray, IntersectionInfo& info) const;
 
     void SetMaterial(Material* material);
     Material* GetMaterial() const;
 
     void Translate(const Vector3d& direction);
-    void Rotate(const Vector3d& axis, float angle);
+    void Rotate(const Vector3d& axis, double angle);
 
     void AddToScene(Scene& scene);
     void Save(Bytestream& stream) const;
@@ -35,7 +35,7 @@ private:
 
     Vector3d pos_;
     Vector3d x_, y_, z_;
-    float length_, radius_;
+    double length_, radius_;
 
     Vector3d invMatU_, invMatV_, invMatW_;
 };
